move duplicate detector to trackerless-network from utils
