+++ packages/trackerless-network/src/logic/RandomGraphNode.ts
@@ -19,7 +19,7 @@ import { PeerList } from './PeerList'
 import { NetworkRpcClient } from '../proto/packages/trackerless-network/protos/NetworkRpc.client'
 import { RemoteRandomGraphNode } from './RemoteRandomGraphNode'
 import { INetworkRpc } from '../proto/packages/trackerless-network/protos/NetworkRpc.server'
-import { DuplicateMessageDetector, NumberPair } from '@streamr/utils'
+import { DuplicateMessageDetector } from './DuplicateMessageDetector'
 import { Logger } from '@streamr/utils'
 import { toProtoRpcClient } from '@streamr/proto-rpc'
 import { IHandshaker } from './neighbor-discovery/Handshaker'
@@ -30,6 +30,7 @@ import { PeerIDKey } from '@streamr/dht/dist/src/helpers/PeerID'
 import { StreamNodeServer } from './StreamNodeServer'
 import { IStreamNode } from './IStreamNode'
 import { ProxyStreamConnectionServer } from './proxy/ProxyStreamConnectionServer'
+import { markAndCheckDuplicate } from './utils'
 
 export interface Events {
     message: (message: StreamMessage) => void
@@ -80,7 +81,7 @@ export class RandomGraphNode extends EventEmitter<Events> implements IStreamNode
             ownPeerDescriptor: this.config.ownPeerDescriptor,
             randomGraphId: this.config.randomGraphId,
             rpcCommunicator: this.config.rpcCommunicator,
-            markAndCheckDuplicate: (msg: MessageRef, prev?: MessageRef) => this.markAndCheckDuplicate(msg, prev),
+            markAndCheckDuplicate: (msg: MessageRef, prev?: MessageRef) => markAndCheckDuplicate(this.duplicateDetectors, msg, prev),
             broadcast: (message: StreamMessage, previousPeer?: string) => this.broadcast(message, previousPeer),
             onLeaveNotice: (notice: LeaveStreamNotice) => {
                 const senderId = notice.senderId
@@ -247,7 +248,7 @@ export class RandomGraphNode extends EventEmitter<Events> implements IStreamNode
 
     broadcast(msg: StreamMessage, previousPeer?: string): void {
         if (!previousPeer) {
-            this.markAndCheckDuplicate(msg.messageRef!, msg.previousMessageRef)
+            markAndCheckDuplicate(this.duplicateDetectors, msg.messageRef!, msg.previousMessageRef)
         }
         this.emit('message', msg)
         
@@ -265,17 +266,6 @@ export class RandomGraphNode extends EventEmitter<Events> implements IStreamNode
             }
         }
         return propagationTargets
-    }
-
-    private markAndCheckDuplicate(currentMessageRef: MessageRef, previousMessageRef?: MessageRef): boolean {
-        const previousNumberPair = previousMessageRef ?
-            new NumberPair(Number(previousMessageRef!.timestamp), previousMessageRef!.sequenceNumber)
-            : null
-        const currentNumberPair = new NumberPair(Number(currentMessageRef.timestamp), currentMessageRef.sequenceNumber)
-        if (!this.duplicateDetectors.has(currentMessageRef.messageChainId)) {
-            this.duplicateDetectors.set(currentMessageRef.messageChainId, new DuplicateMessageDetector(10000))
-        }
-        return this.duplicateDetectors.get(currentMessageRef.messageChainId)!.markAndCheck(previousNumberPair, currentNumberPair)
     }
 
     getOwnStringId(): PeerIDKey {
