+++ packages/trackerless-network/src/logic/proxy/ProxyStreamConnectionClient.ts
@@ -11,7 +11,8 @@ import { IStreamNode } from "../IStreamNode"
 import { EventEmitter } from 'eventemitter3'
 import { ConnectionLocker } from "@streamr/dht/src/exports"
 import { StreamNodeServer } from "../StreamNodeServer"
-import { DuplicateMessageDetector, Logger, NumberPair, wait } from "@streamr/utils"
+import { Logger, wait } from "@streamr/utils"
+import { DuplicateMessageDetector } from "../DuplicateMessageDetector"
 import { PeerList } from "../PeerList"
 import { Propagation } from "../propagation/Propagation"
 import { sampleSize } from 'lodash'
@@ -19,6 +20,7 @@ import { RemoteProxyServer } from "./RemoteProxyServer"
 import { NetworkRpcClient, ProxyConnectionRpcClient } from "../../proto/packages/trackerless-network/protos/NetworkRpc.client"
 import { toProtoRpcClient } from "@streamr/proto-rpc"
 import { RemoteRandomGraphNode } from "../RemoteRandomGraphNode"
+import { markAndCheckDuplicate } from "../utils"
 
 export const retry = async <T>(task: () => Promise<T>, description: string, abortSignal: AbortSignal, delay = 10000): Promise<T> => {
     // eslint-disable-next-line no-constant-condition
@@ -73,7 +75,7 @@ export class ProxyStreamConnectionClient extends EventEmitter implements IStream
         this.server = new StreamNodeServer({
             ownPeerDescriptor: this.config.ownPeerDescriptor,
             randomGraphId: this.config.streamPartId,
-            markAndCheckDuplicate: (msg: MessageRef, prev?: MessageRef) => this.markAndCheckDuplicate(msg, prev),
+            markAndCheckDuplicate: (msg: MessageRef, prev?: MessageRef) => markAndCheckDuplicate(this.duplicateDetectors, msg, prev),
             broadcast: (message: StreamMessage, previousPeer?: string) => this.broadcast(message, previousPeer),
             onLeaveNotice: (notice: LeaveStreamNotice) => {
                 const senderId = notice.senderId
@@ -103,17 +105,6 @@ export class ProxyStreamConnectionClient extends EventEmitter implements IStream
             (msg: StreamMessage, context) => this.server.sendData(msg, context))
         this.rpcCommunicator.registerRpcNotification(LeaveStreamNotice, 'leaveStreamNotice',
             (req: LeaveStreamNotice, context) => this.server.leaveStreamNotice(req, context))
-    }
-
-    private markAndCheckDuplicate(currentMessageRef: MessageRef, previousMessageRef?: MessageRef): boolean {
-        const previousNumberPair = previousMessageRef ?
-            new NumberPair(Number(previousMessageRef!.timestamp), previousMessageRef!.sequenceNumber)
-            : null
-        const currentNumberPair = new NumberPair(Number(currentMessageRef.timestamp), currentMessageRef.sequenceNumber)
-        if (!this.duplicateDetectors.has(currentMessageRef.messageChainId)) {
-            this.duplicateDetectors.set(currentMessageRef.messageChainId, new DuplicateMessageDetector(10000))
-        }
-        return this.duplicateDetectors.get(currentMessageRef.messageChainId)!.markAndCheck(previousNumberPair, currentNumberPair)
     }
 
     async setProxies(
@@ -212,7 +203,7 @@ export class ProxyStreamConnectionClient extends EventEmitter implements IStream
 
     broadcast(msg: StreamMessage, previousPeer?: string): void {
         if (!previousPeer) {
-            this.markAndCheckDuplicate(msg.messageRef!, msg.previousMessageRef)
+            markAndCheckDuplicate(this.duplicateDetectors, msg.messageRef!, msg.previousMessageRef)
         }
         this.emit('message', msg)
         this.propagation.feedUnseenMessage(msg, this.targetNeighbors.getStringIds(), previousPeer ?? null)
