+++ packages/trackerless-network/test/unit/DuplicateMessageDetector.test.ts
@@ -1,4 +1,4 @@
-import { DuplicateMessageDetector, NumberPair, GapMisMatchError, InvalidNumberingError } from '../src/DuplicateMessageDetector'
+import { DuplicateMessageDetector, NumberPair, GapMisMatchError, InvalidNumberingError } from '../../src/logic/DuplicateMessageDetector'
 
 test('starts empty', () => {
     const detector = new DuplicateMessageDetector()
