+++ packages/utils/test/NumberPair.test.ts
@@ -1,4 +1,4 @@
-import { NumberPair } from '../src/DuplicateMessageDetector'
+import { NumberPair } from '../../trackerless-network/src/logic/DuplicateMessageDetector'
 
 test('equalTo', () => {
     expect(new NumberPair(5, 2).equalTo(new NumberPair(5, 3))).toEqual(false)
